+++ pages/sdc_housing.php
@@ -139,12 +139,13 @@ scratch. This page gets rid of all links and provides the needed markup only.
                   <th>Alta</th>
                   <th>Evidencia</th>
                   <th>Observaciones</th>
+                  <th>Modalidad</th>
                   <th width="100">Acciones</th>
                 </tr>
                 </thead>
                 <tbody>
 					<?php
-					$query = "SELECT H.id, H.id_cliente, H.m2, H.sala, H.fila, H.rack, H.observaciones, H.energia, H.fecha_alta, H.evidencia, C.razon_social as cliente, O.razon_social as organismo
+					$query = "SELECT H.id, H.id_cliente, H.m2, H.sala, H.fila, H.rack, H.observaciones, H.energia, H.fecha_alta, H.evidencia, H.modalidad, C.razon_social as cliente, O.razon_social as organismo
                                 FROM sdc_housing as H
                                 INNER JOIN cdc_cliente as C ON H.id_cliente = C.id
                                 LEFT JOIN cdc_organismo as O ON C.id_organismo = O.id
@@ -168,7 +169,14 @@ scratch. This page gets rid of all links and provides the needed markup only.
 							echo '<td align="center">'. $row['rack'].'</td>';
 							echo '<td align="center">'. $row['fecha_alta'].'</td>';
 							echo '<td align="center">'. $row['evidencia'].'</td>';
-							echo '<td align="center">'. $row['observaciones'].'</td>';
+                            echo '<td align="center">'. $row['observaciones'].'</td>';
+                            if ($row['modalidad'] == 1) {
+                                echo '<td align="center">Sala</td>';
+                            } elseif ($row['modalidad'] == 2) {
+                                echo '<td align="center">Rack</td>';
+                            } else {
+                                echo '<td align="center"></td>';
+                            }
               echo '<td align="center">';
               if ($rq_sec['admin']=='1' OR $rq_sec['admin_cli_dc']=='1'){ 
                 echo '<a 
@@ -182,6 +190,7 @@ scratch. This page gets rid of all links and provides the needed markup only.
                   data-energia="' . $row['energia'] . '" 
                   data-observaciones="' . $row['observaciones'] . '" 
                   data-cliente="' . $row['id_cliente'] . '" 
+                  data-modalidad="' . $row['modalidad'] . '" 
                   title="Editar Servicio" class="modal-abm-housing-btn-edit btn btn-sm"><i class="glyphicon glyphicon-edit"></i></a>
                 <a href="sdc_housing.php?aksi=delete&nik='.$row['id'].'" title="Borrar Servicio" onclick="return confirm(\'Esta seguro de borrar el servicio de Housing?\')" class="btn btn-sm"><i class="glyphicon glyphicon-trash"></i></a>';
               }